vim: Add Go syntax and autoformatting
Go used to ship with scripts for vim, but they were removed from the
tree in favor of having the community manage them. The hub for Go
integration in vim has since become [faith/vim-go][].

The autoformatting code in autoload/go was taken from vim-go.

The syntax script was taken from the original Go source. There is a
modified version in vim-go, but i prefer the original, more minimal
version.

[faith/vim-go]: https://github.com/fatih/vim-go/tree/a0d769dd90321765923532f59aa7117449209b9d