move env variables from zshenv to profile
These environment variables define important global settings that need
to be set as early as possible during login, since they impact
potentially every program.

Move them to profile (which should be symlinked from ~/.profile)
because this file is read by many different shells (and even some
desktop environments), and so should increase the chances of it being
loaded.

It turns out zsh is an exception but we can fix that by sourcing the
file manually.